remove value_string.go
